Use normalized nexthops as key in resource accountant
Summary: Normalized nexthops is what we use for key in hwswitch ecmp manager. Resource accountant should also use normalized nexthops as key since it is tracking hw resource usage. This becomes important in NSF because we have topology information in regular nexthop which is not used in hw key and resource accountant will compute may end up not sharing ecmps when topology information is present.

Reviewed By: zechengh09

Differential Revision: D77396971

fbshipit-source-id: ec307c5811b533ca4903b8b23a31131f0bd8da4d