Update DscpMarkingtest to also test DSCP reclassification
Summary:
Modify Dscpmarking test to also test for dscp reclassification feature.

In setup, we create a new dscp reclassification acl which matches NCNF dscp and remaps it to a ICP dscp value.

In verify, we send 2 packets with NCNF dscp value through front panel port and ensure that the number of ICP dscp packets match the amount of NCNF dscp packets sent

Reviewed By: daiwei1983

Differential Revision: D77188505

fbshipit-source-id: 5ce8f804d12361d23c4f9a8aedf0a3a54d927874