fix hw in pause discards counter
Summary:
This patch fixes HwInPauseDiscardsCounterTest.rxPauseEnabled and HwInPauseDiscardsCounterTest.rxPauseDisabled.

For HwInPauseDiscardsCounterTest.rxPauseEnabled, when RX pause is enabled and a pause frame is received, the `In Pause` counter is incremented but the `In discards Raw` counter is not. Therefore for Chenab we mark HwAsic::Feature::IN_PAUSE_INCREMENTS_DISCARDS as unsupported. The test now passes.

For HwInPauseDiscardsCounterTest.rxPauseDisabled, when RX pause is disabled and a pause frame is received, the `In Pause` and `In discards Raw` counters are not incremented. For Chenab, we count this frame against `SAI_PORT_STAT_DOT3_CONTROL_IN_UNKNOWN_OPCODES`. This patch changes the `expectedPktCount` in the test to match this behaviour in the same way as Ebro and Yuba. The test now passes.

Reviewed By: maxwindiff

Differential Revision: D77611109

fbshipit-source-id: 480f72ac5fe308a5d5d105133dd1b3ec99b0de69