Implement persistent file log for fw_util
Summary:
**fw_util Persistent File Log Implementation**

### Summary

This diff implements a persistent file log for the `fw_util` tool. The changes include:

* Adding the `folly/logging:file_handler_factory` dependency to the `BUCK` file.
* Creating a log folder if it doesn't exist in the `fw_util.cpp` file.
* Registering the file handler with the `folly::LoggerDB` to enable logging to the file.

The log folder is set to `/var/facebook/logs/fboss` and the log file name is `fw_util.log`. The changes aim to provide a persistent log for the `fw_util` tool, allowing for easier debugging and troubleshooting.

Reviewed By: somasun

Differential Revision: D76005141

fbshipit-source-id: 43808d3d544a078c2913f17df624d87d9adbf2eb