Add gflag for CDB Command Timeout
Summary:
We noticed in https://fb.workplace.com/groups/fboss.push.investigations/permalink/2138758429961028/, that sometimes `fboss_interface_checker` fails to reach qsfp_service and attempts to call `getTransceiverInfo` are failing to reach. The common scenario we've identified for qsfp_service that causes this is P1864769411.

We suspect that this increased CDB timeout for FR4 lite in D75634749 could be causing the issue. Because FR4-Lite is still in NPI, we are defaulting back to the old timeout, and adding a gflag to support increasing timeout when running on chassis with FR4-Lite modules installed.

Reviewed By: harshitgulati18

Differential Revision: D78031046

fbshipit-source-id: 9305c661313c878fee4dca86fada2edb4c25c838