Add onboarding 'Manuals' section
Summary:
We want to revamp the onboarding checklist by making it easier to follow. It currently provides a broad overview of what needs to get done, but is missing some context, detailed commands, or knowledge we might consider common sense but a vendor won't know. We have a lot of information in our wiki, and these improvements will glue them together in a way that is easily digestible. Our goal is to enable a vendor to follow this checklist from top to bottom and bring up a new platform without asking for help.

This diff:
- creates a new `Manuals` category which contains step by step guides on how to do FBOSS related tasks; gives clarity to respective checklist bullets
- adds some clarity to the first step in the onboarding checklist
- makes the general build and packaging instructions easy to follow from top down like an instruction manual; more organized rather than random info
- changes the order and words used in the nav and sidebar to be consistent

Reviewed By: harshitgulati18

Differential Revision: D77743534

fbshipit-source-id: b4a7243f639bf5c7d09e0075fcbb48278902e313