Use common test port for Flowlet tests
Summary:
This addresses an issue where ACL double counts CPU ingress and front panel count packets.
After this change, the test ACL entries will include SRC_PORT as a match criteria

Lastly, this addresses the flakiness seen with VerifyEcmpRandomSpray test. The test for equality in one case and GE in the other case. It fails sometimes because the equality is miscalculated due to the async nature of packets being sent from CPU.
```
aclPacketCounter(test-udf-flowlet_acl-stats): 390488 -> 400000
```

Differential Revision: D78175970

fbshipit-source-id: ac40cfa2e1bc334c7ed02282154915743898133c