build runtime config
Summary:
In the same way as in the python implementation, build the RuntimeConfig from the platform_manager.json and BspTestConfig

The complicated part here is just in traversing the platform_manager config to know which "actual" adapter is being referenced in each PM Unit. I think the unit tests added here is the best way to understand the problem, but put simply:

If a bus name is listed as "INCOMING@0", in order to actually create and test this device I need to know the actual/real adapter information this is referring to. When I see "INCOMING" in a bus name, I traverse back up through the PM Units to find which slot has an outgoing config to the current unit - this can happen repeatedly.

Reviewed By: somasun

Differential Revision: D75636841

fbshipit-source-id: 0946328dd396f07476daf08b3fb68b8bc0c93e0b