Add a new derived counter to keep track of the num times we hit SRAM buffer low limit
Summary:
We have the SRAM min buffer watermark counter tracking the lowest value seen. But
its good to have a new counter which tells us how many times SRAM buffer availability
dropped below the lowest acceptable value as this will result in PFC on all ports of the
core and hence bad. This counter can be used to decide if we need to add additional
protection for SRAM buffers.

Reviewed By: zechengh09

Differential Revision: D78341244

fbshipit-source-id: c74ea54fe40659cef4b78fdf2e8d1a91b5ca0976