Increment the new counter if SRAM min buffer watermark is lower than low threshold
Summary:
We have the SRAM min buffer watermark counter tracking the lowest value seen. But
its good to have a new counter which tells us how many times SRAM buffer availability
dropped below the lowest acceptable value as this will result in PFC on all ports of the
core and hence bad. This counter can be used to decide if we need to add additional
protection for SRAM buffers.

Reviewed By: maxwindiff

Differential Revision: D78365499

fbshipit-source-id: 79f01cf335ac59ec85187900cbd560aa3662bcf3