introduce expected error handling
Summary:
Some platforms are going to have failures which cannot be reasonably resolved, especially ones built before testing was readily available. A main problem seen so far is that some i2c devices do not appear in i2cdetect even though they are attached to the bus and are creatable.

This introduces an ExpectedError system which allows vendors to define problems which are expected to occur and additionally provide a stated reason. This is similar to PlatformManager's expected errors.

The goal will always be to minimize expected errors, but this will allow us some flexibility.

Reviewed By: RomanChoporov-Meta

Differential Revision: D78426146

fbshipit-source-id: 666d5ae09ea833333e6a28539165bddf282c840a