Generalize early return check in decRouteUsageCount function
Summary:
Early return if no merged groups exist. Can be due to compression
threshold being 0 or if there is a single ECMP groups yet (so
nothing to merge with). If so we return early w/o
adjusting compression penalties. This also allows us
to do a stricter check near the end. This group id
must be present in one of candidateMergeGroups_ or
mergedGroups_ data structures

Differential Revision: D78711547

fbshipit-source-id: 1d440d652c9813baea7ec5303a6dcd5ebdeddef3