Stop high frequency stats thread in BcmSwitch through stopServices
Summary:
Stop high frequency stats thread in BcmSwitch through stopServices. The thread may still be running when the switch is stopped in prod or when a test exits. It solves the following cases:
* The switch is stopped in production
* AgentHwTest in mono mode for any of the following combinations
  * Coldboot and warmboot
  * Setting up and not setting up for warm boot
  * Test successes and failures

`SwSwitchAgentInitializer::stopAgent` is called in `AgentEnsemble::gracefulExit` if setting up for warmboot and also `~MonoAgentEnsemble`. `SwAgentInitializer::stopServices` is called from `SwAgentInitializer::stopAgent` in both the not setting up for warmboot case and through `MonolithicAgentInitializer::handleExitSignal` in the setting up for warmboot case. Since `stopServices` is common, make it virtual allow it to be overridden and call additional stop services, particularly high frequency stats, beforehand.

Previous attempts were made to stop the high frequency stats thread from the destructor and graceful exit, but the destructors did not work because the order of destruction caused issues. The high frequency stats thread depends on `Platform` and `Asic`. Particularly when destroying a `HwAgent` with a `Tomahawk4Platform`, the `Tomahawk4Platform` `std::unique_ptr<Tomahawk4Asic> asic_` member variable is destroyed before the `WedgePlatform` `std::unique_ptr<BcmSwitch> hw_` member variable is destroyed, yielding errors.

Reviewed By: nivinl, maxwindiff

Differential Revision: D78715097

fbshipit-source-id: d57c199156711802cc765ca22c7849565d02c794