Add class to handle user info
Summary:
**Summary**

This diff introduces a new class to handle user information in the Fboss Agent. The changes are made in the `known_bad_test_report.py` file.

**Key Changes**
----------------

*   Imported additional modules such as `getpass`, `os`, `employee`, `async_send_internal_email`, and `OncallThriftClient` to support the new functionality.
*   Updated the `typing` imports to include `Dict` and replaced the `Any` type hint with more specific types where possible.

**Impact**
----------

These changes lay the foundation for a new class that will handle user information, potentially enhancing the overall functionality of the Fboss Agent.

Reviewed By: maxwindiff

Differential Revision: D79107964

fbshipit-source-id: 2f182bf7121ade27ee04925536f89ba83ddd92ce