Add known_bad_test_report.py in buck for execution
Summary:
This diff introduces a new Python library and binary to support the execution of the `known_bad_test_report.py` script.

This diff enables the execution of the `known_bad_test_report.py` script, which is expected to report on known bad tests. The introduction of this script and its dependencies is anticipated to improve the overall testing and reporting infrastructure within the Fboss ecosystem.

Reviewed By: maxwindiff

Differential Revision: D79117177

fbshipit-source-id: 6710626e926bd39e72b53e1cf3909fabd773a088