send report user or oncall
Summary:
## Context

The high-level intention of this stack of diffs is to design, implement, and integrate a new system for reporting known bad tests in the FBOSS Agent.

The diffs progressively build a script, `known_bad_test_report`, which queries Scuba for results of known bad tests. The script is designed to be executable and configurable via command line arguments, allowing for flexible usage and debugging.

Additionally, the diffs introduce a new class to handle user information.

The overall purpose of these diffs is to establish a foundation for automated reporting and management of known bad tests, which will likely improve the testing infrastructure, reduce manual effort, and increase overall efficiency within the FBOSS agent.

The introduction of this system demonstrates a proactive approach to addressing testing challenges and is expected to have a positive impact on the development and maintenance of the Fboss platform.

## This Diff

This diff updates the `known_bad_test_report.py` script to send reports to users or oncalls based on CLI args.

TODO: create task and assign to user or fboss_agent_push oncall based on  CLI args

Reviewed By: maxwindiff

Differential Revision: D79117345

fbshipit-source-id: f0fc229faf2ca63980db183e48fb1ddd370f3cfc