Create task for user or fboss_agent_push oncall
Summary:
## Context

This stack of diffs aims to design and implement a new system for reporting and managing known bad tests in the FBOSS Agent. The goal is to automate the process of identifying and reporting known bad tests, reducing manual effort, and increasing overall efficiency in the testing infrastructure.

The diffs progressively build a Python script, `known_bad_test_report`, which queries Scuba for results of known bad tests and can be executed and configured via command-line arguments. The script is designed to be flexible, allowing for easy usage and debugging. Additionally, the diffs introduce a new class to handle user information, laying the foundation for sending reports to users or oncalls based on CLI arguments.

The overall purpose of these diffs is to establish a foundation for automated reporting and management of known bad tests, which will likely improve the testing infrastructure, reduce manual effort, and increase overall efficiency within the FBOSS agent. The introduction of this system demonstrates a proactive approach to addressing testing challenges and is expected to have a positive impact on the development and maintenance of the Fboss platform.

## This Diff

This diff creates a task for a user or the fboss_agent_push oncall.

Differential Revision: D79274455

fbshipit-source-id: cfa6af59ef03a902c9e8273899ea7d1e021b7b1f