Modify the default list attr size in replayer to be 4K
Summary:
With the new vendor switch attr creation as part of DNX SAI12.2, we are using almost 1400 entries and thats
exceeding the current default. This results in truncated list which fails if replayer is used by vendor. To avoid
failures for vendors, change the default list size to be large enough to accomodate create_vendor_switch().

There are 6 lists created by default, so the overall utilization change would be negligible.

Reviewed By: zechengh09, maxwindiff

Differential Revision: D79416282

fbshipit-source-id: a6018b5505e42b2161314b2908b55d548df660aa