Add agent high frequency stats test PFCTest
Summary: Add an AgentHwTest for high frequency stats collection. The test starts a high frequency stats collection job collecting stats every 20ms for 200ms with a thrift client while simultaneously running a thread to send PFC frames. This test verifies that PFC counters are collected using a thrift client to get the high frequency stats APIs on the hw agent. Positivity of the counters are verified to guarantee that the PFC counter for each port and PG ID where PFC frames were sent are positive.

Reviewed By: nivinl

Differential Revision: D78526250

fbshipit-source-id: 675157382b31d2af5cf3d790cb652312e2a6ac67