Add total known bad to email and task
Summary: This diff modifies the `known_bad_test_report.py` file to include the total number of known bad tests in the email and task notifications. The changes include updating the `HTTP_format_known_bad_list` method to accept an additional `known_bad_tests` parameter and modifying the method to include the total count of known bad tests in the HTML output.

Differential Revision: D79541749

fbshipit-source-id: 2cb32facf07952846101c06ae63ad99b9d4ae606