Remove FbossEepromInterface::getFieldDictionary() method
Summary:
## This Diff

### Summary
This diff removes the `getFieldDictionary()` method from the `FbossEepromInterface` class. The method was previously used to return a map of EEPROM field entries, but it is no longer needed and has been removed to simplify the codebase.

The change consists of two parts:

* The `getFieldDictionary()` method is removed from the `FbossEepromInterface.cpp` file.
* The declaration of the `getFieldDictionary()` method is removed from the `FbossEepromInterface.h` file.

This change does not affect the functionality of the code, but rather simplifies the interface and reduces the amount of code that needs to be maintained.

Reviewed By: Scott8440

Differential Revision: D78359750

fbshipit-source-id: d4a54ee2c50249e3e825e1f5a6967ffca7026f4c