Add AgentHostPrefixRouteScaleTest
Summary:
In S543220, morgan800cc hits route scale issue due to:

Host prefix (/128) are saved in TCAM (MSB) and SRAM (LSB) hybrid mode. And we got many new BAG ips which changing on high bits then causing TCAM entry exhausted.

We  disabled router counter + added new SDK patch D79258262 to reduce the length of MSB in TCAM so to increase the scalability.
This diff is to add a test case for regression.

In this test case, 10X /128 are generated based on the BAG IP pattern
```
2401:db00:e233:59:6600:0:d:0/128
2401:db00:e233:59:6800:0:d:0/128
2401:db00:e233:59:6a00:0:d:0/128
2401:db00:e233:59:6c00:0:d:0/128
```
Also 200 /128 changing on bits 20 to simulate different region / location.
```
::/0 [Generic] root
    2401:db00::/32 [RIR] Organization:   ORG-FSPL3-AP
        2401:db00:e200::/40 [Region] US_EAST - Interconnect
            2401:db00:e230::/44 [Metro] NAO - Interconnect
                2401:db00:e233:59:6c00:0:d:0/128 [NetworkDevice] bag183.nao3
```

Reviewed By: zechengh09

Differential Revision: D79761534

fbshipit-source-id: d197542da7741048b09cca63e210581bf7634bbb