Append to high frequency stats on positive watermarks
Summary: It's possible for watermark stats to be the same between two polls, but it indicates that watermarks are present for both the timestamps and the intervals on which they were measured because they're cleared. Append high frequency stats on positive watermarks as well as stat changes for PFC counters.

Reviewed By: nivinl

Differential Revision: D79920485

fbshipit-source-id: 78504e914ab78351b89e641eda9b6a1faaa8e773