Skipping LoadBalancing ECMP tests in ProdInvariantTests for Yamp platforms
Summary: Skipping ECMP Tests and DSCPtoQueueMapping Tests in ProdInvariantTests for FA devices as it contains portchannels. In some configs, the number of ports which are not members of portchannels and have an ip assigned are less than kEcmpWidth. When we encounter such configs, we skip the ecmp and DSCP to queue mapping tests as we dont have enough number of ECMP ports.

Reviewed By: zechengh09

Differential Revision:
D79209353

Privacy Context Container: L1297311

fbshipit-source-id: 48e01fbba84d707e0040b997fd88a394dcb03aae