Adding functions to determine uplinks downlinks for FA devices in Invariant Tests
Summary: Skipping ECMP Tests and DSCPtoQueueMapping Tests in ProdInvariantTests for FA devices as it contains portchannels. In some configs, the number of ports which are not members of portchannels and have an ip assigned are less than kEcmpWidth. When we encounter such configs, we skip the ecmp and DSCP to queue mapping tests as we dont have enough number of ECMP ports.

Reviewed By: zechengh09

Differential Revision: D79107769

fbshipit-source-id: 8d6c5e9511e1893e70ba2f9493c6986504db9b3a