Add SAI test, BSP tests, and gtest filters for ease of use
Summary:
This diff adds T0, T1, and T2 SAI tests as well as a BSP tests section. We also provide `--gtest_filter` flags for easy copy and paste.

Tests are added as conf files that can be used with `run_test.py`. In order to enable this and reference the files at code snippets, a symlink titled `fboss` is placed in the `github` directory. The shipit mapping has been updated to ignore this symlink in this diff: D80013501

Reviewed By: harshitgulati18

Differential Revision: D79860872

fbshipit-source-id: 3ceace2b38dd1cfb6bf91656dedc7c9f8cc4d6a4