Adjust assert to accomodate EcmpResourceMgr in merge groups mode
Summary:
If ecmpDemandExceeded and we have backupEcmpGroupType_ set,
then this group should spillover to backup ecmpType config.
For merge group mode, we are not guaranteed which groups
will be picked for merge, so we can't assert that the
new groups  will always be of merge type.

Differential Revision:
D79848296

Privacy Context Container: L1297311

fbshipit-source-id: b8116500273084a3d4f3490863d24078f2329766