Create reusable code snippets
Summary:
We are defining reusable code snippets so that we can have our manuals be self-contained. We can reuse code snippets in multiple manuals to prevent readers from having to jump to different pages, while also avoiding duplication. It also makes our life easier if we want to change some commands as they will apply to all pages using the snippets.

This diff:
- creates code snippets
- edits build doc to use the code snippets
- edits packaging doc to use the code snippets

Reviewed By: joseph5wu

Differential Revision: D79570278

fbshipit-source-id: 9a050af81cc4eb96864c1f4231ece2c71999f1b3