Update and add new manuals
Summary:
This diff should provide the steps and guidance vendors need to follow to complete step 1 of the onboarding checklist. Mainly, adding Platform Manager and BSP support. And some guidance on building, running, and testing platform services.

We are utilizing code snippets defined in D79570278 so that we can avoid duplicating code blocks. This is useful because we want our manuals to be self contained; vendors shouldn't have to jump to many different pages and should just need one manual to achieve a task. This requires lots of duplicate code, which is perfect for the code snippets.

Here's a list of what this diff achieves:
- adds a new manual which handles step 1 of the onboarding checklist
- adds a troubleshooting guide
- edit onboarding checklist to include new manual and troubleshooting guide
- edit onboarding checklist to reflect outcomes of the new manual
- edit navbar "Manuals" to point to the new manual
- edit PlatformManager doc to add a note linking the thrift file in our repo
- delete obsolete manual
- minor clarity improvements to some code snippets

Reviewed By: joseph5wu

Differential Revision: D79847437

fbshipit-source-id: 44b90262d633985b285e2c501ca461a384f6852a