Fix Lane Map and Polarity swap generation for XGS
Summary:
For XGS devices, the lane map is in this format:

RX_LANE_MAP=0x20317645

The nibble is indexed from the connector lane number perspective. The value in each nibble is the mapped lane.  For example, if RX lane 1 is mapped to connector lane 3, then nibble 3 is set to 1.

In our logic, we were doing reverse. We considered the index to be physical lane, and the nibble value to be logical lane.

Fixed the generating logic in this diff, and also corrected the CSVs for TH5 devices accordingly. We had incorrect CSVs for TH5 to match the incorrect chip config generation logic.

Reviewed By: huruinan

Differential Revision: D80218230

fbshipit-source-id: 21acd159d0d00072fd9d454f5404ceb317d10f43