fw_util: fw_util does not support setting gpioChipValue according to configuration (#497)
Summary:
# Description

The current version of the fboss fw_util only allows setting the `gpioChipValue` to 1 through hardcoding and does not accept a value of 0 from the configuration.

<img width="743" height="246" alt="image" src="https://github.com/user-attachments/assets/97196be3-eca1-41ef-869e-dda438c3e977" />

We need to update the code to support setting the `gpioChipValue` based on the configuration in `fw_util.json`.

# Motivation

The fw_util tool does not support setting `gpioChipValue` to 0 as specified in the configuration file `fw_util.json`. This limitation caused the pic_cpld upgrade to fail in the Icetea project because it expected `gpiochip0` to be 0, according to the config value in `fw_util.json`.

The issue also affects the pic_cpld upgrade in the Icecube project.

Pull Request resolved: https://github.com/facebook/fboss/pull/497

Test Plan:
1. **Reproduce the issue:** Set `"gpioChipValue": "0"` in `fw_util.json` for pic_cpld. However, the code runs with `gpiochip0` set to 1, causing the test case to fail.
<img width="736" height="504" alt="image" src="https://github.com/user-attachments/assets/8bd2ff92-b4ab-43df-b0dd-8c3664f598e2" />
<img width="1885" height="770" alt="image" src="https://github.com/user-attachments/assets/0339d414-33f3-44fa-9a52-870f52a07475" />

2. **Run fw_util binary with the fixed patch:** The test case works successfully after applying the fix.
<img width="973" height="504" alt="image" src="https://github.com/user-attachments/assets/3a6d9b6d-fd64-4c83-af2e-f4308115fb6f" />

3. **Review detailed logs:**
[fw_util_reproduce_log.txt](https://github.com/user-attachments/files/21404656/fw_util_reproduce_log.txt)
[fw_util_fix_log.txt](https://github.com/user-attachments/files/21404659/fw_util_fix_log.txt)

Reviewed By: Scott8440

Differential Revision: D79414388

Pulled By: mikechoifb

fbshipit-source-id: 67e8d4f1d7eeb1cb0e44f90fd3d370cd58d951ab