Fix bug in sFlow PktFactory allocation and add unit test
Summary:
The sFlow PktFactory construction logic has a bug where the allocation for the packet may be smaller than needed. Fix the allocation logic because a multiple of `XDR_BASIC_BLOCK_SIZE` is needed.

Add a unit test for this coverage. If the fix is not there, the code throws an `std::out_of_range` `overflow` error from IOBuf.

Reviewed By: maxwindiff

Differential Revision: D80306722

fbshipit-source-id: 7eaf688f8778afa7ae95ad331db4318cfd869ef1