Add logic for deciding which merged groups to reclaim
Summary:
Logic is
- Start from the highest cost to lowest cost groups
- Calculate unmerge cost, if it fits in the available space
add it to ordered list of groups to unmerge.
- Find all the other groups that are part of the merge set
and add them to the ordered list as well. E.g. say we have
a merge set of groups [1, 3] and [2, 4]. Both of which have
a cost of say X. Then in our ordered (by cost) group we may see
the groups in the following order [4, 3, 2, 1]. And say we have
space to create 2 more groups. If we go in the above order
order at step where we unmerge 3, we will have prefixes pointing
to the following groups
- groups  4, 3 and the original merged groups [1, 3] and [2, 4].
At this point we already created 2 more groups so we can't proceed.
- OTOH if we unmerge merged groups together. Viz. we go in the order
(say) [4, 2, 3, 1] - we will be able to unmerge both and free
up the merged groups. This also follows that we merge in unmerge
a merge set together.

Reviewed By: ashutoshgrewal, msomasundaran

Differential Revision:
D80056109

Privacy Context Container: L1297311

fbshipit-source-id: fa30821d7c2348dbb5a317d22283d7795ee6798a