Add file size check to prevent crash
Summary:
A hardware issue in a device was causing all i2c reads to fail, including an i2c eeprom. That caused an exception because `tellg` was returning -1. Just check for that and return an exception that PlatformManager can catch.

resize is marked noexcept or something like that causing it to immediately terminate rather than propagate the error

Reviewed By: RomanChoporov-Meta

Differential Revision: D80203580

fbshipit-source-id: cca555f679ce8887d59545f72aadff45d809e47a