Add schedule as ci.diff
Summary:
We've seen a lot of "false triggers" of platform mapping sync jobs to configerator in the past 1-2 months. There is two problems here:
1. It's easy to forget to commandeer / land synced configerator changes (these diffs do not auto-publish right now). The solution to this should come in one of the next diffs.
2. The `ci_sandcastle` macro was not defined correctly. We assumed it should only run when certain targets in the specified changeset are modified (https://fburl.com/code/6jv50519). However, it doesn't seem like this solution works when this macro runs on continuous runs (e.g. https://fburl.com/chronos/hznnzrzx).

This diff aims to solve problem 2 by restricting this macro to only run at diff time, which should significantly reduce these false triggers. It seems like the default behavior for not defining the `schedules` field of `ci_sandcastle` macro is to trigger on all schedules, instead of just diff time.

Reviewed By: kevin645

Differential Revision: D80262722

fbshipit-source-id: eb709e7eb2c3d7288cfd18bbfb84512965b84647