Re-land "Guard counter reading behind SAI_PORT_PG_DROP_STATUS"
Summary: The code in question is guarded behind `#if defined(BRCM_SAI_SDK_GTE_13_0)` as well as `HwAsic::Feature::SAI_PORT_PG_DROP_STATUS` which is off on Ramon, so it's unlikely to be the root cause of 12.2 R3 test failures (https://fburl.com/conveyor/9w6jn6m8). Ran the test 20 times as well.

Reviewed By: zechengh09

Differential Revision: D80478812

fbshipit-source-id: 31f7430d6bbd0a6fdfad6439a7eb423deb5b5204