Make interface case-insensitive
Summary:
A recent change to the darwin config file broke all fw_util actions except for `list`. The casing of the target names was changed to uppercase, while logic in fw_util uses lowercased strings to find the fpd.

Arista noticed this breakage and submitted a PR: https://github.com/facebook/fboss/pull/490 but I thought it would be better if we handled this in a general way instead of adding a special case for certain platforms.

What this change does is make the CLI interface case-insensitive for `--fw_target_name`.

The output is unaffected, meaning we use the exact string in the config.

Reviewed By: joancaneus

Differential Revision: D80472311

fbshipit-source-id: 73b2985f0b0efe5e8a161e76497cc2923ed62426