deprecate AgentCoppTest.VerifyCoppPpsLowPri
Summary:
AgentCoppTest.VerifyCoppPpsLowPri is marked as known bad globally
https://www.internalfb.com/code/configerator/[c77dd73b6754f52e6e885540331135c4899523b0]/source/neteng/netcastle/test_config/sai_agent_test.cinc?lines=21

However, the common known bad list is not picked up by Morgan800cc so it is the only one running it.  **(But it will also fail for morgan800cc once it moves to ASAN mode)**

The test is trying to inject 60k packets (measure the time-A of all injection) which will punt to CPU low-priv queue. Then verify low-pri pps by (CPU low-pri count /  time-A).

This is flaky and not working once the expectd low-pri pps > the inject-up pps from CPU to ASIC, and the inject-up pps from CPU to ASIC was affected by many things like ASAN or SW->HW thrift call.

So do not think we shoudl try to fix this, we should try to enable  this one (unfortunately also disabled globally for now, but it makes more sense as it " Create dataplane loop with lowerPriority traffic on port0"):
AgentCoppQueueStuckTest.CpuQueueHighRateTraffic
https://www.internalfb.com/code/fbsource/[ad12206a0d3a36c010d55e55c5d69e648c548b9a]/fbcode/fboss/agent/test/agent_hw_tests/AgentCoppTests.cpp?lines=1780

**Things to follow-up:**

1. try to fix and enable AgentCoppQueueStuckTest.CpuQueueHighRateTraffic to test low priv queue PPS/BPS

2. Morgan800cc low-priv queue rate limit does not seem like working. Need to debug with Cisco

Reviewed By: vasantspatil

Differential Revision: D80489422

fbshipit-source-id: 4171b471d171ba3898eafaba543c513d5c496c35