Return bool instead of void for send packet APIs
Summary:
### Summary of Changes

This diff makes changes to the return type of several packet sending APIs in the Fboss Agent codebase. Specifically, the following APIs now return a boolean value instead of void:

* `sendNetworkControlPacketAsync`
* `sendPacketAsync`
* `sendPacketOutOfPortAsync`

The changes aim to provide a way to indicate whether the packet sending operation was successful or not. The boolean return type allows the caller to check the status of the operation and take appropriate action.

Reviewed By: msomasundaran

Differential Revision: D80523283

fbshipit-source-id: 569d07f99860c32ef6c49a2e64dd5f1d93e6134d