Define flag in LACP TransmitMachine
Summary:
### Summary

Doc: S549441[ LACP RX Timer Expire in FA](https://docs.google.com/document/d/1LbUtX6rlRUjY1n72OW42RYbHFdUcvQqjRNGp1nciC3w/edit?tab=t.0#heading=h.pm8hm85sdpy7)

This diff defines a new flag `isLastTransmissionSuccessful` in the LACP TransmitMachine to track whether the last transmission was successful. The flag is updated in the `LacpMachines.cpp` file to set it to `false` when the transmission fails.

The purpose of this diff is to provide a way to track the success of the last LACP transmission.

Reviewed By: msomasundaran

Differential Revision: D80523282

fbshipit-source-id: cd1f5df8b17f6bbbc225a46db77e2533610d14f9