Add MAC TX queue stuck counter to fb303 stats
Summary:
In S545783, we noticed an issue with RCI stuck and to debug that, need to add min/max
watermarks for MAC TX data queue. The goal is to eventually monitor these watermarks to
check if they are changing or are stuck. A stuck MAC TX queue will result in RCI stuck case
as well.

Reviewed By: ashutoshgrewal

Differential Revision: D80721168

fbshipit-source-id: 79a143f5657e108c4ea09df31e6af7d22b0abdde