Thrift support for MAC transmit data queue min/max watermarks
Summary:
In S545783, we noticed an issue with RCI stuck and to debug that, need to add min/max
watermarks for MAC TX data queue. The goal is to eventually monitor these watermarks to
check if they are changing or are stuck. A stuck MAC TX queue will result in RCI stuck case
as well.

Reviewed By: maxwindiff

Differential Revision: D80557276

fbshipit-source-id: b955fffa7cdfc75d1b3e5c3ec332c623fe194222