Add a new feature for MAC transmit data queue watermark support
Summary:
In S545783, we noticed an issue with RCI stuck and to debug that, need to add min/max
watermarks for MAC TX data queue. The goal is to eventually monitor these watermarks to
check if they are changing or are stuck. A stuck MAC TX queue will result in RCI stuck case
as well.

Reviewed By: maxwindiff

Differential Revision: D80557280

fbshipit-source-id: c40d4e0768415b9643208540b0aa9eaa0854367a