Set TX MAC data queue stuck based on MAC queue min/max watermarks
Summary:
In S545783, we noticed an issue with RCI stuck and to debug that, need to add min/max
watermarks for MAC TX data queue. The goal is to eventually monitor these watermarks to
check if they are changing or are stuck. A stuck MAC TX queue will result in RCI stuck case
as well.

Reviewed By: maxwindiff

Differential Revision: D80674857

fbshipit-source-id: 970cf7552c9ec90ba7d8e59c2c6822e896d2f41b