Add a new port level counter to indicate if the FIFO is stuck or not
Summary:
In S545783, we noticed an issue with RCI stuck and to debug that, need to add min/max
watermarks for MAC TX data queue. The goal is to eventually monitor these watermarks to
check if they are changing or are stuck. A stuck MAC TX queue will result in RCI stuck case
as well.

Reviewed By: maxwindiff

Differential Revision: D80674855

fbshipit-source-id: fb2655bc233ca7e5bc158161fccb579fddc2b36e