Collect and update the MAC tx data queue min/max watermarks
Summary:
In S545783, we noticed an issue with RCI stuck and to debug that, need to add min/max
watermarks for MAC TX data queue. The goal is to eventually monitor these watermarks to
check if they are changing or are stuck. A stuck MAC TX queue will result in RCI stuck case
as well.

Reviewed By: maxwindiff

Differential Revision: D80557275

fbshipit-source-id: e70a79be28819ecf58eddd8c4a22e82744498572