Fix calculation for how reclaimed groups contribute to the remaining space
Summary:
We were calculating space remaining as
```
(total)canReclaim - reclaimedGroups.size()
```
But this is not correct. Consider a reclaimed merge group of size 2.
This creates 2 primary groups and removes a merged (primary) group.
That means that the incremental demand is of 1 group. While if we
just look at reclaimedGroups.size() it would subtract 2 from remaining
space.
So we change this calculation to a running calculation, where after
reclaiming any merge set, we update spaceRemaining as
```
spaceRemaining -= mergeSet.size() - 1

Reviewed By: ashutoshgrewal

Differential Revision:
D80498908

Privacy Context Container: L1297311

fbshipit-source-id: 85331644a7eb308cc771c103b76cdf5b642e3c95