darwin: Add PEM_EEPROM symlink pm config (#521)
Summary:
weutil hw test fails on darwin as the PEM EEPROM is defined in the weutil config, but there is no corresponding symlink. This PR fixes that by editing the PM config so that it will manually generate the appropriate symlink.

Tested that PM runs successfully, weutil hw test passes.

Pull Request resolved: https://github.com/facebook/fboss/pull/521

Reviewed By: Scott8440

Differential Revision: D80856476

Pulled By: somasun

fbshipit-source-id: b18273854624769e662748242baa2fdbbc86be4e