Improve the accuracy of DataStore::hasPmUnit
Summary:
The hasPmUnit() function is meant to return whether a PmUnit is
present on the slot or not. After a lot of changes in DataStore, this function
currently returns true even if we have detected a PmUnit is absent in the slot.

Refer to inline documentation of the function for details of this change

Reviewed By: Scott8440

Differential Revision: D80990051

fbshipit-source-id: 0e8688efb6baf1c9bb342729d08df4af2d911d28