Fix a bug for warm boots
Summary:
The application of AEC depends on the host interface code, and not the media interface code. Doing the right comparison now in setApplicationCodeLocked

Previously the AEC cable was detecting a different applicable programmed because we were checking the mediaInterface instead of hostInterface code, so it was setting the application again which flaps the link.

Reviewed By: shiva-menta

Differential Revision: D80521598

fbshipit-source-id: bf77ad1867edcec1d5c7243a900c08550f31c898