Update checks in mirror programming for bcm agent mirror tests
Summary:
`isMirrorProgrammed` returns false if the BCM mirror is programmed so update the logic to return false if the mirror is *not* programmed.

A tunnel need not be present on a mirror for a port span mirror ([`addMirrorConfig`](https://www.internalfb.com/code/fbsource/[6d257008c348]/fbcode/fboss/agent/test/utils/MirrorTestUtils.cpp?lines=59-72) logic) so return true in cases where there might not be a tunnel.

Reviewed By: maxwindiff, nivinl

Differential Revision: D81056630

fbshipit-source-id: f5d26f57f857d8f2e234c5ad68a17f166a168132