Fixed CreateMaxUcmpMembers Testcase for TH3
Summary:
Currently getUcmpMembersAndWeight() does not assign weights to ECMP nexthop members considering the maxVariableWIdthECMPSize. It just assign weights, 2,3,2,3.... for all next hop groups until the total number of ECMP members hits kMaxUcmpMembers. This is a problem for ASICs like TH3 which has the maxVariableWidthEcmpSize to be 128.  If the total weight exceeds 128, it resorts to wide ECMP where the weights are normalized to a number that is power of 2 (256,512).
In this diff, we make sure the total weight of each UCMP Nexthop group is below the maxVariableWidthEcmpSize. This way we can exhaust all the allowed kMaxUcmpMembers and also adhere to total weight of each next hop group being less than maxVariableWidthEcmpSize.

Reviewed By: zechengh09

Differential Revision:
D80776063

Privacy Context Container: L1297311

fbshipit-source-id: e85f7a53dc07bc37b2e779a74cfff485d0d59635