Add Event Queueing / Handling Functions to Port Manager
Summary: As titled - these are the same exact functions we use in TransceiverManager.cpp for enqueueing / managing events. Making a copy of these functions for use in PortManager.h. I was thinking to try to use some form of inheritance here so we don't need to copy these over, but given the few number of functions and difference in class constructor args, I opted for just copying.

Reviewed By: harshitgulati18

Differential Revision: D77256077

fbshipit-source-id: 6ad45e6936e828af05b3a75ab815ecfc7e55873c