When deleting route, look at deleted route from latest state delta
Summary:
Look the removed route from previous state. Its possible that
we may have cleared override nhops info in that state as part
of a previous unmerge. If so we would have accounted for that
as a new primary ECMP group and we need to account for its
removal here.

For e.g. consider merge group [1, 2] pointing to prefixes P1 and P2.
Where before the merge we had P1->G1, P2->G2.
Now imagine we get at route delete for both P1 and P2 in the same update.
In this case when we delete P1, we would delete merge group and restore
P2->G2 mapping. Now this is a primary ECMP group. When we get a delete
for P2, which in turn deletes G2, we should decrement the primary ecmp
group count. But if we look at state of P2 from original state, it will
still be pointing to a merged group. Leading us to not decrement this count.

Reviewed By: msomasundaran

Differential Revision:
D80594288

Privacy Context Container: L1297311

fbshipit-source-id: 47e095a735934aef0479ea4319dc687038a79a04